Custom Elements management added

+++ src/migrations/2017_01_01_000006_create_mediables_table.php
@@ -8,17 +8,20 @@ class CreateMediablesTable extends Migration
 {
     public function up()
     {
-        Schema::create('mediables', function (Blueprint $table) {
-            $table->integer('media_id')->unsigned()->nullable();
-            
-            $table->integer('mediable_id')->unsigned()->nullable();
-            $table->string('mediable_type')->nullable()->default(null);
+        if (!Schema::hasTable('mediables')) {
+            Schema::create('mediables', function (Blueprint $table) {
+                $table->integer('media_id')->unsigned()->nullable();
+                
+                $table->integer('mediable_id')->unsigned()->nullable();
+                $table->string('mediable_type')->nullable()->default(null);
 
-            $table->foreign('media_id')->references('id')->on('medias')->onDelete('cascade');
+                $table->foreign('media_id')->references('id')->on('medias')->onDelete('cascade');
 
-            $table->integer('position')->unsigned()->default(0);
-            $table->string('collection')->nullable()->default(null);
-        });
+                $table->integer('position')->unsigned()->default(0);
+                $table->string('collection')->nullable()->default(null);
+                $table->string('table')->nullable()->default(null);
+            });
+        }
     }
 
     public function down()
